WIP: removed wildcard support
