+++ components/token/CollectionInfo.tsx
@@ -1,6 +1,5 @@
 import useCollection from 'hooks/useCollection'
 import useDetails from 'hooks/useDetails'
-import getMode from 'lib/getMode'
 import { optimizeImage } from 'lib/optmizeImage'
 import Link from 'next/link'
 import React, { FC } from 'react'
@@ -8,10 +7,9 @@ import React, { FC } from 'react'
 type Props = {
   collection: ReturnType<typeof useCollection>
   details: ReturnType<typeof useDetails>
-  mode: ReturnType<typeof getMode>['mode']
 }
 
-const CollectionInfo: FC<Props> = ({ collection, mode, details }) => {
+const CollectionInfo: FC<Props> = ({ collection, details }) => {
   const token = details.data?.tokens?.[0]
 
   const tokenDescription =
@@ -21,13 +19,7 @@ const CollectionInfo: FC<Props> = ({ collection, mode, details }) => {
   return (
     <article className="col-span-full rounded-2xl border border-gray-300 bg-white p-6 dark:border-neutral-600 dark:bg-black">
       <div className="reservoir-h5 mb-4 dark:text-white">Collection Info</div>
-      <Link
-        href={
-          mode === 'collection'
-            ? '/'
-            : `/collections/${collection.data?.collection?.id}`
-        }
-      >
+      <Link href={`/collections/${collection.data?.collection?.id}`}>
         <a className="inline-flex items-center gap-2">
           <img
             src={optimizeImage(
