+++ hooks/useCollections.ts
@@ -9,7 +9,7 @@ const PROXY_API_BASE = process.env.NEXT_PUBLIC_PROXY_API_BASE
 
 type Collections = paths['/collections/v2']['get']['responses']['200']['schema']
 
-export default function useCollections() {
+export default function useCollections(fallback?: Collections) {
   const { ref, inView } = useInView()
 
   const pathname = `${PROXY_API_BASE}/collections/v2`
@@ -19,6 +19,7 @@ export default function useCollections() {
     fetcher,
     {
       revalidateFirstPage: false,
+      fallback,
     }
   )
