+++ hooks/useUserTokens.ts
@@ -13,7 +13,6 @@ type Tokens =
 export default function useUserTokens(
   collectionId: string | undefined,
   fallbackData: Tokens[],
-  mode: string | undefined,
   user: string | undefined
 ) {
   const { ref, inView } = useInView()
@@ -23,7 +22,7 @@ export default function useUserTokens(
   const tokens = useSWRInfinite<Tokens>(
     (index, previousPageData) =>
       getKey(
-        { pathname, mode, collectionId, proxyApi: PROXY_API_BASE },
+        { pathname, collectionId, proxyApi: PROXY_API_BASE },
         index,
         previousPageData
       ),
@@ -48,7 +47,6 @@ type InfiniteKeyLoader = (
   custom: {
     pathname: string
     collectionId: string | undefined
-    mode: string | undefined
     proxyApi: string | undefined
   },
   ...base: Parameters<SWRInfiniteKeyLoader>
@@ -58,13 +56,12 @@ const getKey: InfiniteKeyLoader = (
   custom: {
     pathname: string
     collectionId: string | undefined
-    mode: string | undefined
     proxyApi: string | undefined
   },
   index: number,
   previousPageData: paths['/users/{user}/tokens/v2']['get']['responses']['200']['schema']
 ) => {
-  const { pathname, collectionId, mode, proxyApi } = custom
+  const { pathname, collectionId, proxyApi } = custom
   if (!proxyApi) {
     console.debug(
       'Environment variable NEXT_PUBLIC_PROXY_API_BASE is undefined.'
@@ -78,14 +75,6 @@ const getKey: InfiniteKeyLoader = (
   let query: paths['/users/{user}/tokens/v2']['get']['parameters']['query'] = {
     limit: 20,
     offset: index * 20,
-  }
-
-  if (mode === 'community') {
-    query.community = collectionId
-  }
-
-  if (mode === 'collection') {
-    query.collection = collectionId
   }
 
   const href = setParams(pathname, query)
